Moved startCheckmatePractice() from guipractice to checkmatepractice
